+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Text alignment".
 - Tests for "Elliptic curve cryptography".
 - Tests for "Train passenger".
+- Tests for "CGS minifier".
 
 ### Fixed
 - Tests groups for "CGFunge interpreter".